// lib/supabase/database.types.ts
export type Database = {
  public: {
// ...
    Enums: {
      app_permission: "users.manage" | "requests.approve" | "items.manage";
      app_role: "member" | "admin" | "approver";
      request_status: "submitted" | "approved" | "rejected";
    };
    CompositeTypes: {
      [_ in never]: never;
    };
  };
};